+++ src/main/java/insane96mcp/survivalreimagined/utils/Utils.java
@@ -1,5 +1,7 @@
 package insane96mcp.survivalreimagined.utils;
 
+import com.ezylang.evalex.Expression;
+import com.ezylang.evalex.data.EvaluationValue;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
@@ -74,5 +76,26 @@ public class Utils {
         random.setSeed(random.nextLong());
         random.setSeed(random.nextLong());
         return random;
+    }
+
+    /**
+     * Returns the hp regenerated each second
+     */
+    public static float computeFoodFormula(FoodProperties food, String formula) {
+        Expression expression = new Expression(formula);
+        try {
+            //noinspection ConstantConditions
+            EvaluationValue result = expression
+                    .with("hunger", food.getNutrition())
+                    .and("saturation_modifier", food.getSaturationModifier())
+                    .and("effectiveness", getFoodEffectiveness(food))
+                    .and("fast_food", food.isFastFood())
+                    .evaluate();
+            return result.getNumberValue().floatValue();
+        }
+        catch (Exception ex) {
+            LogHelper.error("Failed to evaluate food formula: %s", expression);
+            return -1f;
+        }
     }
 }